Move everything related to IIDM and Groovy to IIDM scripting module
