--- iidm/iidm-api/src/main/groovy/com/powsybl/iidm/network/SubstationExtension.groovy
@@ -1,20 +1,0 @@
-/**
- * Copyright (c) 2019, RTE (http://www.rte-france.com)
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/.
- */
-
-package com.powsybl.iidm.network
-
-/**
- *
- * @author Geoffroy Jamgotchian <geoffroy.jamgotchian at rte-france.com>
- */
-class SubstationExtension {
-
-    static Object getCountry(Substation self) {
-        // forward the getCounty method to getNullableCountry
-        self.getNullableCountry()
-    }
-}
