--- iidm/iidm-api/src/main/resources/META-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,3 +1,0 @@
-moduleName = powsybl-iidm-module
-moduleVersion = 1.0
-extensionClasses = com.powsybl.iidm.network.IdentifiableExtension  com.powsybl.iidm.network.NetworkExtension com.powsybl.iidm.network.SubstationExtension
