+++ iidm/iidm-scripting/src/main/resources/META-INF/groovy/org.codehaus.groovy.runtime.ExtensionModule
@@ -1,0 +1,3 @@
+moduleName = powsybl-iidm-scripting-module
+moduleVersion = 1.0
+extensionClasses = com.powsybl.iidm.network.scripting.IdentifiableExtension  com.powsybl.iidm.network.scripting.NetworkExtension com.powsybl.iidm.network.scripting.SubstationExtension
