+++ loadflow/loadflow-validation/pom.xml
@@ -41,6 +41,10 @@
     <dependencies>
         <!-- Compilation dependencies -->
         <dependency>
+            <groupId>org.codehaus.groovy</groupId>
+            <artifactId>groovy</artifactId>
+        </dependency>
+        <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>powsybl-iidm-api</artifactId>
             <version>${project.version}</version>
